requirements.txt
# Document ingestion
pymupdf==1.24.1
Pillow==10.3.0
pdf2image==1.17.0

# OCR
paddleocr
paddlepaddle>=3.0.0

# Layout detection
transformers==4.40.0
torch==2.3.0
torchvision==0.18.0

# Markdown processing
markdown-it-py==3.0.0

# Backend
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0

# Utilities
numpy==1.26.4
opencv-python==4.9.0.80